Refactored remote-server. Improved prometheus logic.
