+++ remote-server/api-server/taskNodeMapping.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
-	db "github.com/scarpart/distributed-task-scheduler/db/sqlc"
+	db "github.com/scarpart/distributed-task-scheduler/remote-server/db/sqlc"
 )
 
 type CreateMappingRequest struct {
